plugins/bark-gpu/forward/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
PyYAML>=5.4.0
requests>=2.26.0
aiohttp>=3.8.0
